service/pyproject.toml
[project]
name = "service"
version = "1.6.6"
description = "The backend service of Xyzen"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "aioboto3>=15.5.0",
    "alembic>=1.16.4",
    "anthropic>=0.67.0",
    "celery[redis]>=5.4.0",
    "croniter>=3.0.0",
    "databases[postgresql,sqlite]>=0.9.0",
    "fastapi[standard]>=0.116.1",
    "fastmcp>=2.12.3",
    "google-genai>=1.38.0",
    "greenlet>=3.2.3",
    "httpx>=0.28.1",
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langchain-google-genai>=3.0.0",
    "langchain-google-vertexai>=3.0.3",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.0",
    "llm-sandbox[docker,k8s]>=0.3.20",
    "loguru>=0.7.3",
    "mcp>=1.11.0",
    "novu-py>=3.13.0",
    "openai>=1.98.0",
    "psycopg[binary,pool]>=3.2.9",
    "py-vapid>=1.9.0",
    "pywebpush>=2.0.0",
    "pydantic-graph>=0.4.10",
    "pydantic-settings>=2.10.1",
    "pyjwt[crypto]>=2.10.0",
    "pymupdf>=1.26.7",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sqlmodel>=0.0.24",
    "openpyxl>=3.1.2",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "websockets>=13.0",
    "reportlab>=4.4.7",
    "langchain-qwq>=0.3.1",
    "jinja2>=3.1.6",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "pyyaml>=6.0.3",
    "pytesseract>=0.3.13",
    "pillow>=12.0.0",
    "celery-types>=0.24.0",
    "trafilatura>=1.12.0",
    "daytona>=0.139.0",
    "e2b-code-interpreter>=1.0.0",
    "langmem>=0.0.30",
    "openfga-sdk>=0.9.0",
    "dashscope>=1.25.12",
]

[dependency-groups]
dev = [
    "types-pyjwt>=1.7.1",
    "types-requests>=2.32.4.20250611",
    "types-aioboto3>=15.5.0",
    "types-openpyxl>=3.1.5.20250919",
    "pre-commit>=4.2.0",
    "ruff>=0.14.3",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.28.1",
    "aiosqlite>=0.20.0",
    "pytest-xdist>=3.8.0",
    "polyfactory>=3.2.0",
    "watchdog>=3.0.0",
    "basedpyright>=1.38.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
known-first-party = [
    "app",
    "core",
    "handler",
    "repo",
    "models",
    "schemas",
    "utils",
    "common",
    "internal",
    "middleware",
]

# --- basedpyright: used by CI, pre-commit, and `just type-backend` ---
# Pylance (VS Code) config is in .vscode/settings.example.json
[tool.basedpyright]
include = ["."]
exclude = ["**/migrations/**", "**/.venv/**", "**/tests/**"]
pythonVersion = "3.13"
pythonPlatform = "All"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

# Downgrade "Unknown" type noise from incomplete 3rd-party stubs
reportUnknownMemberType = "hint"
reportUnknownParameterType = "hint"
reportUnknownArgumentType = "hint"
reportUnknownVariableType = "hint"
reportUnknownLambdaType = "hint"

# Downgrade rules too strict for LangChain/LangGraph ecosystem
reportMissingParameterType = "hint"
reportMissingTypeArgument = "hint"
reportUnnecessaryIsInstance = "hint"
reportUnnecessaryComparison = "hint"
reportUntypedFunctionDecorator = "hint"
reportUntypedClassDecorator = "hint"
reportUntypedBaseClass = "hint"
reportAttributeAccessIssue = "hint"

# Keep visible as warnings
reportUnnecessaryTypeIgnoreComment = "warning"
reportPrivateUsage = "warning"

# Not useful for this codebase
reportMissingTypeStubs = "none"


[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=85",
    "-v",
    "--tb=short",
    "--maxfail=10",
    "--durations=10",
    "--color=yes",
    # "-n", "auto",  # Enable parallel execution with pytest-xdist (uncomment to use)
    "--dist", "worksteal",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "migrations/*",
    "**/__pycache__/*",
    ".venv/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
